+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # 1.1
 - Fixed year
-- Adjusted `newdrive` for verification on if a drive already exists
+- Adjusted `newdrive` for verification on if a drive already exists or uses invalid characters
 - Added basic `jpac` and `pacman` responses (may be edited in the future)
 - Added start scripts